fetch/metadata/generated/js-module-via-script-cross-domain.tentative.sub.html


<!-- This test was procedurally generated. Please do not modify it directly. -->
<html lang="en">
  <meta encoding="utf-8">
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/fetch/metadata/resources/helper.js"></script>
  <script type="module"
    src="http://{{domains[www2]}}:{{ports[http][0]}}/fetch/metadata/resources/record-header.py?file=script-{{$id:uuid()}}"></script>
  <body>
  <script>
  'use strict';
  promise_test(function() {
    return pollForRequestRecording('script-{{$id}}')
      .then(function(actual) {
var expected = {
  dest: 'script',
  mode: '',
  site: '',
  user: ''
};
assert_header_equals(actual, expected);

      });
  }, 'something across domains');
  </script>
</html>
